requirements.txt
aif360==0.3.0
altair==4.1.0
numpy==1.18.5
pandas==1.0.4
scikit-learn==0.23.1
seaborn==0.10.1
shap==0.35.0
streamlit==0.61.0
PyYAML==5.3.1
